requirements.txt
Django == 2.0.1
djangorestframework == 3.7.7
markdown == 2.6.11
Pillow == 5.0.0